% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addLegend.R
\name{addLegend}
\alias{addLegend}
\title{Add a Legend to a Plot}
\usage{
addLegend(spectra, leg.loc, use.sym, ...)
}
\arguments{
\item{spectra}{An object of S3 class \code{\link[ChemoSpec]{Spectra}} or 
\code{\link[ChemoSpec2D]{Spectra2D}}.}

\item{leg.loc}{Character; Any string acceptable to \code{\link{legend}}.}

\item{use.sym}{Logical; if true, the color scheme is set to black and the
points plotted with symbols.  Applies only to \code{ChemoSpec}.}

\item{\dots}{Additional parameters to be passed to the plotting functions.}
}
\value{
None.  Side effect is to modify an existing plot.
}
\description{
Add a legend to a plot created by either \code{ChemoSpec} or \code{ChemoSpec2D}.
\code{ChemoSpec2D} does not use symbols.
}
\examples{
if (requireNamespace("ChemoSpec", quietly = TRUE)) {
  library("ChemoSpec")
  data(SrE.IR)
  plotSpectra(SrE.IR, main = "SrE.IR Data",
   which = c(1, 5, 16), offset = 0.1, yrange = c(0, 0.5))
  addLegend(SrE.IR, use.sym = FALSE, leg.loc = list(x = 3500, y = 0.5), bty = "n") 
}

}
\author{
Bryan A. Hanson, DePauw University.
}
