% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check4Gaps.R
\name{check4Gaps}
\alias{check4Gaps}
\title{Check for Discontinuities (Gaps) in a Vector & Optionally Make a Plot}
\usage{
check4Gaps(x, tol = 0.01, y = NULL, silent = FALSE, ...)
}
\arguments{
\item{x}{A numeric vector to be checked for gaps.}

\item{tol}{A number indicating the tolerance for checking to see if the step
between successive \code{x} values are the same.  Depending upon how the
\code{x} values are stored and rounded, you may need to change the value of
\code{tol} to avoid flagging trivial "gaps".}

\item{y}{An optional vector of \code{y}-values which correspond to the \code{x}
values.  Only used in \code{ChemoSpec}.  If provided, a plot will be made
in the style of a \code{\link[ChemoSpec]{Spectra}} object showing the gap(s).}

\item{silent}{Logical indicating a "no gap" message
should not be reported to the console.  Important because
\code{check4Gaps} is called iteratively by other functions.}

\item{\dots}{Other parameters to be passed to the plot routines if
\code{y} is provided, e.g. \code{xlim}.}
}
\value{
A data frame giving the data chunks found, with one chunk per row.
Also a plot if {y} is provided.  In the event there are no gaps found,
a data frame with zero rows is returned.  The data frame has columns as follows:
\item{beg.freq }{The first frequency value in a given data chunk.}
\item{end.freq }{The last frequency value in a given data chunk.}
\item{size }{The length (in frequency units) of the data chunk.}
\item{beg.indx }{The index of the first frequency value in the data chunk.}
\item{end.indx }{The index of the last frequency value in the data chunk.}
}
\description{
The basic procedure is to compare x[n + 1] - x[n] for successive values of
n.  When this value jumps, there is a gap which is flagged. \code{beg.indx}
and \code{end.indx} will always be contiguous as indices must be; it is the
\code{x} values that jump or have the gap.  The indices are provided as they
are more convenient in some programming contexts.  If not assigned, the
result appears at the console.
}
\examples{

x <- seq(from = 5, to = 12, by = 0.1)
remove <- c(8:11, 40:45); x <- x[-remove]
gaps <- check4Gaps(x)

if (requireNamespace("ChemoSpec", quietly = TRUE)) {
  library("ChemoSpec")
  data(SrE.IR)
  newIR <- removeFreq(SrE.IR, rem.freq = SrE.IR$freq > 1800 & SrE.IR$freq < 2500)
  check4Gaps(x = newIR$freq, y = newIR$data[1,])
}

if (requireNamespace("ChemoSpec2D", quietly = TRUE)) {
  library("ChemoSpec2D")
  data(MUD1)

  plotSpectra2D(MUD1, which = 7, lvls = seq(-1, 1, by = 0.2),
    main = "MUD1 Sample 7: Complete Data Set")

  MUD1a <- removeFreq(MUD1, remF2 = 6 ~ high, remF1 = 13 ~ 17)
  # Cannot plot this data, results would be misleading.
  sumSpectra(MUD1a) # sumSpectra calls check4Gaps for each dimension
}

}
\author{
Bryan A. Hanson, DePauw University.
}
\keyword{utilities}
