% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChemoSpecUtils-package.R
\docType{package}
\name{ChemoSpecUtils-package}
\alias{ChemoSpecUtils-package}
\alias{ChemoSpecUtils}
\title{Functions Supporting Packages ChemoSpec and ChemoSpec2D}
\description{
Functions supporting the packages ChemoSpec and ChemoSpec2D.
}
\author{
Bryan A. Hanson (DePauw University), Tejasvi Gupta.

Maintainer: Bryan A. Hanson \email{hanson@depauw.edu}
}
