% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chkGraphicsOpt.R
\name{chkGraphicsOpt}
\alias{chkGraphicsOpt}
\title{Check the Graphic Output Option/Mode}
\usage{
chkGraphicsOpt(silent = TRUE)
}
\arguments{
\item{silent}{Logical.  Silences most messages if \code{TRUE}.}
}
\value{
Character.  The value of the current graphics output option/mode.
}
\description{
This function should be used to check the current graphics mode.  It simply reports
the current setting unless it is invalid in which case the mode is set to \code{base}.
}
\seealso{
See \code{\link{GraphicsOptions}} for more information about the graphics options.
}
\author{
Bryan A. Hanson (DePauw University), Tejasvi Gupta.
}
\keyword{utilities}
