% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addEllipseInfo.R
\name{.addEllipseInfo}
\alias{.addEllipseInfo}
\title{Add Info About Ellipses to Plot}
\usage{
.addEllipseInfo(ellipse)
}
\arguments{
\item{ellipse}{String giving the type of ellipse; one of \code{cls}, \code{rob} or \code{both}.}
}
\value{
None. Side effect is a modified plot.
}
\description{
\emph{Internal} function.
}
\author{
Bryan A. Hanson (DePauw University).
}
\keyword{internal}
