% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addMethod.R
\name{.addMethod}
\alias{.addMethod}
\title{Add Info About Data Reduction Method to Plot}
\usage{
.addMethod(pca)
}
\value{
None.  Side effect is a modifed plot.
}
\description{
\emph{Internal} function.
}
\author{
Bryan A. Hanson (DePauw University).
}
\keyword{internal}
