% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chkArgs.R
\name{.chkArgs}
\alias{.chkArgs}
\title{Check spectra and pca Arguments of Functions That Require Them}
\usage{
.chkArgs(mode = 11L)
}
\arguments{
\item{mode}{Integer code giving the type of check to be run.}
}
\value{
None.  Stops if there is a problem with the arguments.
}
\description{
\emph{Internal} function.
}
\author{
Bryan A. Hanson (DePauw University).
}
\keyword{internal}
