% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chkReqGraphicsPkgs.R
\name{.chkReqGraphicsPkgs}
\alias{.chkReqGraphicsPkgs}
\title{Verify Required Graphics Packages are Available.}
\usage{
.chkReqGraphicsPkgs(pkgs)
}
\arguments{
\item{pkgs}{Character. Vector of package names.  Each is checked for availability.}
}
\value{
\code{NULL}, invisibly.
}
\description{
\emph{Internal function}.
}
\author{
Bryan A. Hanson (DePauw University), Tejasvi Gupta.
}
\keyword{internal}
