% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colLeaf.R
\name{.colLeaf}
\alias{.colLeaf}
\title{Color the Leaves of a Dendrogram Based on a Spectra Object}
\usage{
.colLeaf(n, spectra)
}
\arguments{
\item{n}{A node in a dendrogram object.}

\item{spectra}{An object of S3 class \code{\link[ChemoSpec:Spectra]{ChemoSpec::Spectra()}}.}
}
\value{
Returns a node with the label color properties set.
}
\description{
\emph{Internal function.} This function colors the leaves of a dendrogram object.  The code was taken
from the help files.
}
\author{
Bryan A. Hanson (DePauw University).
}
\keyword{internal}
