% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeEllipses.R
\name{.computeEllipses}
\alias{.computeEllipses}
\title{Compute Confidence Ellipses}
\usage{
.computeEllipses(x, quan = 1/2, alpha = 0.025)
}
\arguments{
\item{x}{As per \code{\link[mvoutlier]{corr.plot}}.}

\item{quan}{As per \code{\link[mvoutlier]{corr.plot}}.}

\item{alpha}{As per \code{\link[mvoutlier]{corr.plot}}.}
}
\value{
A list with the following elements (a simpler version of that in the
original function \code{\link[mvoutlier]{corr.plot}}): \item{x.cls}{The x
values for the classical ellipse.} \item{y.cls}{The y values for the
classical ellipse.} \item{c}{The correlation value for the classical
ellipse.} \item{x.rob}{The x values for the robust ellipse.}
\item{y.rob}{The y values for the robust ellipse.} \item{r}{The correlation
value for the robust ellipse.}
}
\description{
\emph{Internal function}. A utility function which when given a \code{x,y} data set computes both classical
and robust confidence ellipses.
}
\seealso{
See function \code{\link[mvoutlier]{corr.plot}} in package
\pkg{mvoutlier} on which this function is based.
}
\author{
Bryan A. Hanson (DePauw University).
}
\keyword{internal}
