% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distPlot.R
\name{.distPlot}
\alias{.distPlot}
\title{Draw a Distance Plot}
\usage{
.distPlot(spectra, M, method, ...)
}
\arguments{
\item{spectra}{An object of S3 class \code{\link[ChemoSpec:Spectra]{ChemoSpec::Spectra()}} or \code{\link[ChemoSpec2D:Spectra2D]{ChemoSpec2D::Spectra2D()}}.}

\item{M}{Matrix of class \code{dist} containing the distances.}

\item{method}{String.  The distance method used to create the matrix.}

\item{\dots}{Parameters to be passed to the plotting routines. \emph{Applies to base graphics only}.}
}
\value{
None.  Side effect is a plot.
}
\description{
\emph{Internal function}.
}
\author{
Bryan A. Hanson (DePauw University).
}
\keyword{internal}
