% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawEllipses.R
\name{.drawEllipses}
\alias{.drawEllipses}
\title{Add Ellipses to Plot}
\usage{
.drawEllipses(ELL, gr, ellipse, use.sym, ...)
}
\arguments{
\item{ELL}{Coordinates of the ellipse points.}

\item{gr}{Group information.}

\item{ellipse}{Type of ellipse; one of \code{cls}, \code{rob} or \code{both}.}

\item{use.sym}{Logical.  Should symbols be used?  For this function, determines color of lines.}

\item{\ldots}{Parameters to be passed to the plotting routines. \emph{Applies to base graphics only}.}
}
\value{
None.  Side effect is a modified plot.
}
\description{
\emph{Internal function}.
}
\author{
Bryan A. Hanson (DePauw University).
}
\keyword{internal}
