% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraData.R
\name{.extraData}
\alias{.extraData}
\title{Check for Extra Data Appended to a Spectra or Spectra2D Object}
\usage{
.extraData(spectra)
}
\arguments{
\item{spectra}{An object of S3 class \code{\link[ChemoSpec:Spectra]{ChemoSpec::Spectra()}} or \code{\link[ChemoSpec2D:Spectra2D]{ChemoSpec2D::Spectra2D()}}.}
}
\value{
.  Logical. \code{TRUE} if a mismatch was found for the data lengths.  Message are emited as well.
}
\description{
\emph{Internal function}.
}
\author{
Bryan A. Hanson (DePauw University).
}
\keyword{internal}
