% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findNA.R
\name{.findNA}
\alias{.findNA}
\title{Find NA in a Spectra2D Object}
\usage{
.findNA(spectra, retFreq = FALSE)
}
\arguments{
\item{spectra}{An object of S3 class \code{\link[ChemoSpec2D:Spectra2D]{ChemoSpec2D::Spectra2D()}}.}

\item{retFreq}{Logical. Should the frequencies be returned?}
}
\value{
A list with two elements giving the indices of NAs for rows and columns,
unless \code{retFreq = TRUE}, in which case the frequencies are returned.
}
\description{
\emph{Internal function}. This function identifies the extent of any NA in a \code{\link[ChemoSpec2D:Spectra2D]{ChemoSpec2D::Spectra2D()}} object.
This information can be used for summaries and plots.
}
\author{
Bryan A. Hanson (DePauw University).
}
\keyword{internal}
