% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getExtremeCoords.R
\name{.getExtremeCoords}
\alias{.getExtremeCoords}
\title{getExtremeCoords}
\usage{
.getExtremeCoords(data, names, tol)
}
\arguments{
\item{data}{A matrix containing the x values of the points/samples in the
first column, and the y values in the second.}

\item{names}{A character vector of sample names.  Length must match the
number of rows in \code{data}.}

\item{tol}{\code{.writeDoc_Tol()}}
}
\value{
A list with elements \code{x}, \code{y} and \code{l}, giving the coordinates,
and labels of the most extreme data points.
}
\description{
\emph{Internal function}. A utility function which returns a list of the extreme
coordinates that are to be labelled.
}
\author{
Bryan A. Hanson (DePauw University), Tejasvi Gupta.
}
\keyword{internal}
