% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLimits.R
\name{.getLimits}
\alias{.getLimits}
\title{Process Limits for a Spectra2D Plot}
\usage{
.getLimits(spectra, dim, form)
}
\arguments{
\item{spectra}{An object of S3 class \code{\link[ChemoSpec2D:Spectra2D]{ChemoSpec2D::Spectra2D()}}.}

\item{dim}{The dimension for which the limits are desired.}

\item{form}{A formula specification.}
}
\value{
An object of S3 class \code{\link[ChemoSpec2D:Spectra2D]{ChemoSpec2D::Spectra2D()}}.
}
\description{
\emph{Internal function}. Get limits from a user specified formula
(used in \code{\link[ChemoSpec2D:removeFreq]{ChemoSpec2D::removeFreq()}} and \code{\link[ChemoSpec2D:removePeaks2D]{ChemoSpec2D::removePeaks2D()}}).
The user may not know or think about whether F2 or F1 is ascending or descending
so we will try to get it right no matter how the user gives
the formula; e.g. 6 ~ 3 ought to be handled as 3 ~ 6.
}
\author{
Bryan A. Hanson (DePauw University).
}
\keyword{internal}
