% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVarExplained.R
\name{.getVarExplained}
\alias{.getVarExplained}
\title{Compute the Variance Explained from PCA or Other Reduction Methods}
\usage{
.getVarExplained(so)
}
\arguments{
\item{so}{A object returned by a data reduction method.}
}
\value{
A vector giving the variance explained by each successive component.
}
\description{
\emph{Internal function}.
}
\author{
Bryan A. Hanson (DePauw University).
}
\keyword{internal}
