% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggAnnotate.R
\name{.ggAnnotate}
\alias{.ggAnnotate}
\title{Annotate a ggplot2 Plot with Text}
\usage{
.ggAnnotate(text, x, y, ...)
}
\arguments{
\item{text}{Character.  The text to be plotted, but certain values are intercepted
for special handling.}

\item{x}{Numeric. The x coordinate for the text to be plotted, in NPC.}

\item{y}{Numeric. The y coordinate for the text to be plotted, in NPC.}

\item{\dots}{Additional arguments to be passed to \code{textGrob}.}
}
\value{
A custom annotation grob, ready to be added to an existing plot.
}
\description{
\emph{Internal function}. General purpose function to add text to a ggplot2 plot.  Uses
\code{textGrob} so arguments like \code{hjust} can be passed via ...
}
\author{
Bryan A. Hanson (DePauw University), Tejasvi Gupta.
}
\keyword{internal}
