% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggRepel.R
\name{.ggRepel}
\alias{.ggRepel}
\title{Wrapper for geom_text_repel}
\usage{
.ggRepel(df)
}
\arguments{
\item{df}{Data frame, containing elements \code{x}, \code{y} and \code{label}.}
}
\value{
A geom that can be added to an existing plot.
}
\description{
\emph{Internal function}. Wrapper for \code{geom_text_repel} to make customization, modifcation
and maintenance easy.
}
\author{
Bryan A. Hanson (DePauw University).
}
\keyword{internal}
