% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelExtremes.R
\name{.labelExtremes}
\alias{.labelExtremes}
\title{Label Extreme Values}
\usage{
.labelExtremes(data, names, tol)
}
\arguments{
\item{data}{A matrix containing the x values of the points/samples in the
first column, and the y values in the second.}

\item{names}{A character vector of sample names.  Length must match the
number of rows in \code{x}.}

\item{tol}{\code{.writeDoc_Tol()}}
}
\value{
None.  Annotates the plot with labels.
}
\description{
\emph{Internal function}. A utility function which plots the sample names next to the sample points.
The number of samples labeled can be specified by passing it from the
calling function. An internal function, not generally called by the user.
}
\author{
Bryan A. Hanson (DePauw University), Tejasvi Gupta.
}
\keyword{internal}
