% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHCA.R
\name{.plotHCA}
\alias{.plotHCA}
\title{Plot Dendrogram for Spectra or Spectra2D Object}
\usage{
.plotHCA(spectra, hclst, sub.title, use.sym, leg.loc, ...)
}
\arguments{
\item{spectra}{An object of S3 class \code{\link[ChemoSpec:Spectra]{ChemoSpec::Spectra()}} or \code{\link[ChemoSpec2D:Spectra2D]{ChemoSpec2D::Spectra2D()}}.}

\item{hclst}{A \code{\link{hclust}} object.}

\item{sub.title}{A character string for the subtitle.}

\item{use.sym}{Logical; if true, the color scheme will be black and
lower-case letters will be used to indicate group membership.
Applies to \code{Spectra} objects only.}

\item{leg.loc}{Character; if \code{"none"} no legend will be drawn.
Otherwise, any string acceptable to \code{\link{legend}}.}

\item{\dots}{Parameters to be passed to the plotting routines. \emph{Applies to base graphics only}.}
}
\value{
An object of class \code{\link{dendrogram}}. Side effect is a plot.
}
\description{
\emph{Internal function}. This function plots the results of an HCA analysis.
}
\author{
Bryan A. Hanson (DePauw University).
}
\keyword{internal}
