% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepLegendCoords.R
\name{.prepLegendCoords}
\alias{.prepLegendCoords}
\title{Sanitize Legend Coordinates}
\usage{
.prepLegendCoords(go, leg.loc, x.min = 0, x.max = 0, y.min = 0, y.max = 0)
}
\arguments{
\item{go}{A string specifying the graphics mode.}

\item{leg.loc}{A list giving x and y coordinates.}

\item{x.min}{A number giving the minimum x value in the data. Used only if graphics mode is set to base.}

\item{x.max}{A number giving the maximum x value in the data. Used only if graphics mode is set to base.}

\item{y.min}{A number giving the minimum y value in the data. Used only if graphics mode is set to base.}

\item{y.max}{A number giving the maximum y value in the data. Used only if graphics mode is set to base.}
}
\value{
A list of coordinates specifying x and y position of the legend along with a position
designation, except in one case where a string is returned.
}
\description{
\emph{Internal function}.
}
\author{
Bryan A. Hanson (DePauw University), Tejasvi Gupta.
}
\keyword{internal}
