% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remGrpSam.R
\name{.remGrpSam}
\alias{.remGrpSam}
\title{Remove Groups or Samples from a Spectra or Spectra2D Object}
\usage{
.remGrpSam(spectra, kill, group = TRUE)
}
\arguments{
\item{spectra}{An object of S3 class \code{\link[ChemoSpec:Spectra]{ChemoSpec::Spectra()}} or \code{\link[ChemoSpec2D:Spectra2D]{ChemoSpec2D::Spectra2D()}}.}

\item{kill}{A vector of character strings giving a regex pattern to be applied to the sample
or group names.  If \code{group = FALSE} can also be an integer vector giving the samples
to be removed.}

\item{group}{Logical. \code{TRUE} means we are removing groups.  \code{FALSE} means we are removing samples.}
}
\value{
An object of S3 class \code{\link[ChemoSpec:Spectra]{ChemoSpec::Spectra()}} or \code{\link[ChemoSpec2D:Spectra2D]{ChemoSpec2D::Spectra2D()}}.
}
\description{
\emph{Internal function}.
}
\author{
Bryan A. Hanson (DePauw University).
}
\keyword{internal}
