% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorePlot.R
\name{.scorePlot}
\alias{.scorePlot}
\title{scorePlot}
\usage{
.scorePlot(
  spectra,
  so,
  pcs = c(1, 2),
  ellipse = "none",
  tol = "none",
  use.sym = FALSE,
  leg.loc = "topright",
  ...
)
}
\description{
\emph{Internal function}.  Parameters etc are documented in \code{\link[=plotScores]{plotScores()}}.
}
\author{
Bryan A. Hanson (DePauw University), Tejasvi Gupta.
}
\keyword{internal}
