% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrinkLeaf.R
\name{.shrinkLeaf}
\alias{.shrinkLeaf}
\title{Shrink the Leaves of a Dendrogram Based on a Spectra Object}
\usage{
.shrinkLeaf(n, spectra)
}
\arguments{
\item{n}{A node in a dendrogram object.}

\item{spectra}{An object of S3 class \code{\link[ChemoSpec:Spectra]{ChemoSpec::Spectra()}}.}
}
\value{
Returns a node with the label size properties set.
}
\description{
\emph{Internal function}. This function shrinks the size of leaves of a dendrogram object.  The code
was taken from the help files.
}
\references{
\url{https://github.com/bryanhanson/ChemoSpec}
}
\author{
Bryan A. Hanson (DePauw University).
}
\keyword{internal}
