% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumGroups.R
\name{sumGroups}
\alias{sumGroups}
\title{Summarize the Group Membership of a Spectra or Spectra2D Object}
\usage{
sumGroups(spectra)
}
\arguments{
\item{spectra}{An object of S3 class \code{\link[ChemoSpec]{Spectra}} or
\code{\link[ChemoSpec2D]{Spectra2D}} whose group membership information is desired.}
}
\value{
A data frame as follows. Note that if there are groups with no
members these are dropped.
 \item{group}{The name of the group.}
 \item{no.}{The number in the group.}
 \item{color}{The color assigned to the group.}
 \item{symbol}{The symbol assigned to the group. \code{Spectra} objects only.}
 \item{alt.symbol}{The alternative symbol assigned to the group. \code{Spectra} objects only.}
}
\description{
This function summarizes the group membership	 of
a \code{Spectra} or \code{\link[ChemoSpec2D]{Spectra2D}} object.
}
\examples{
if (checkForPackageWithVersion("ChemoSpec", "5.1")) {
  library("ChemoSpec")
  data(SrE.IR)
  sumGroups(SrE.IR)
}

if (checkForPackageWithVersion("ChemoSpec2D", "0.3")) {
  library("ChemoSpec2D")
  data(MUD1)
  sumGroups(MUD1)
}
}
\seealso{
To summarize the entire object, \code{\link{sumSpectra}}.
}
\author{
Bryan A. Hanson, DePauw University.
}
\keyword{utilities}
