% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addLegend.R
\name{addLegend}
\alias{addLegend}
\title{Add a Legend to a Plot}
\usage{
addLegend(spectra, leg.loc, use.sym, ...)
}
\arguments{
\item{spectra}{An object of S3 class \code{\link[ChemoSpec]{Spectra}} or 
\code{\link[ChemoSpec2D]{Spectra2D}}.}

\item{leg.loc}{Character; Any string acceptable to \code{\link{legend}}.}

\item{use.sym}{Logical; if true, the color scheme is set to black and the
points plotted with symbols.  Applies only to \code{ChemoSpec}.}

\item{\dots}{Additional parameters to be passed to the plotting functions.}
}
\value{
None.  Side effect is to modify an existing plot.
}
\description{
Add a legend to a plot created by either \code{ChemoSpec} or \code{ChemoSpec2D}.
\code{ChemoSpec2D} does not use symbols.
}
\author{
Bryan A. Hanson, DePauw University.
}
